[ISSUE-#16] - EmailSenderService를 Interface로 추상화하고, 프로덕션 환경과 테스트 환경에서 주입되는 구현체 분리 - 통합 테스트 환경에서 @Transactional 어노테이션 전부 제거 - CertificationCode 엔티티 수정   - ttl를 해당 엔티티 내부에서 결정하는 것이 아닌 비즈니스 로직에서 정해주도록 팩토리 메서드 생성자 수정   - CertificationCode 도메인 내부에서 사용되는 상수는 constant 패키지 내부에서 관리   - 통합 테스트에 만료 기간이 지난 인증 코드 검증 요청 시나리오 추가

+++ src/main/java/com/eventty/eventtynextgen/certification/response/CertificationSendCodeResponseView.java
@@ -11,7 +11,7 @@ public class CertificationSendCodeResponseView {
         this.message = message;
     }
 
-    public static CertificationSendCodeResponseView createMessage(String email, long ttl) {
+    public static CertificationSendCodeResponseView createMessage(String email, int ttl) {
         String message = String.format("%s로 인증 코드를 발송했습니다. %d분 안에 인증 코드를 입력해 주세요.", email, ttl);
 
         return new CertificationSendCodeResponseView(message);
